+++ CHANGELOG.md
@@ -9,6 +9,20 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ...
 
+## [0.3.0] - 2015-11-21
+
+### Added
+- Add `Encoder` and `Decoder` protocols to core with stream-based methods
+  `encode!` and `decode!`, respectively.
+- Add `encode` and `decode` to core which operate on byte arrays.
+- Add `multicodec.codecs` namespace with a few useful codec implementations.
+
+### Changed
+- Rename `core/paths` to `core/headers`.
+- Move header functions to new `multicodec.header` namespace.
+- On validation errors, header functions now throw exceptions with type
+  information via `ex-data`.
+
 ## [0.2.0] - 2015-11-19
 
 ### Added
@@ -19,5 +33,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial project release
 
-[Unreleased]: https://github.com/greglook/clj-multicodec/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/greglook/clj-multicodec/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/greglook/clj-multicodec/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/greglook/clj-multicodec/compare/0.1.0...0.2.0